nettoyage du watchdog
