+++ qsequoia2/QSEQUOIA2_dockwidget_base.ui
@@ -6,7 +6,7 @@
    <rect>
     <x>0</x>
     <y>0</y>
-    <width>356</width>
+    <width>676</width>
     <height>699</height>
    </rect>
   </property>
@@ -57,13 +57,6 @@
             </widget>
            </item>
            <item>
-            <widget class="QPushButton" name="watchdog">
-             <property name="toolTip">
-              <string>paramètres watchdog</string>
-             </property>
-            </widget>
-           </item>
-           <item>
             <widget class="QPushButton" name="add_project">
              <property name="toolTip">
               <string>Créer un nouveau projet</string>
@@ -109,7 +102,7 @@
         <rect>
          <x>0</x>
          <y>0</y>
-         <width>332</width>
+         <width>652</width>
          <height>474</height>
         </rect>
        </property>
