--- qsequoia2/scripts/watchdog/watchdog_label.ui
@@ -1,135 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>watchdog_setting</class>
- <widget class="QFrame" name="watchdog_setting">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>404</width>
-    <height>193</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>Kartenn data connect</string>
-  </property>
-  <layout class="QVBoxLayout" name="rootLayout">
-   <item>
-    <widget class="QLabel" name="label">
-     <property name="text">
-      <string>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p align=&quot;center&quot;&gt;Welcome !&lt;/p&gt;&lt;p align=&quot;center&quot;&gt;Watchdog manager: listens to the selected folder for adding layers&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</string>
-     </property>
-     <property name="alignment">
-      <set>Qt::AlignCenter</set>
-     </property>
-    </widget>
-   </item>
-   <item>
-    <widget class="QGroupBox" name="groupBox">
-     <property name="title">
-      <string>connections</string>
-     </property>
-     <layout class="QVBoxLayout" name="connectionsLayout">
-      <item>
-       <layout class="QHBoxLayout" name="autoLayout">
-        <item>
-         <widget class="QLabel" name="label_auto">
-          <property name="text">
-           <string>From the auto mode</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <spacer name="spacer_auto">
-          <property name="orientation">
-           <enum>Qt::Horizontal</enum>
-          </property>
-          <property name="sizeHint" stdset="0">
-           <size>
-            <width>0</width>
-            <height>0</height>
-           </size>
-          </property>
-         </spacer>
-        </item>
-        <item>
-         <widget class="QRadioButton" name="radio_auto">
-          <property name="checked">
-           <bool>true</bool>
-          </property>
-         </widget>
-        </item>
-       </layout>
-      </item>
-      <item>
-       <layout class="QHBoxLayout" name="downloadsLayout">
-        <item>
-         <widget class="QLabel" name="down">
-          <property name="text">
-           <string>From your download folder</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <spacer name="spacer_downloads">
-          <property name="orientation">
-           <enum>Qt::Horizontal</enum>
-          </property>
-          <property name="sizeHint" stdset="0">
-           <size>
-            <width>0</width>
-            <height>0</height>
-           </size>
-          </property>
-         </spacer>
-        </item>
-        <item>
-         <widget class="QRadioButton" name="radio_downloads"/>
-        </item>
-       </layout>
-      </item>
-     </layout>
-    </widget>
-   </item>
-   <item>
-    <layout class="QHBoxLayout" name="okLayout">
-     <item>
-      <spacer name="leftSpacer">
-       <property name="orientation">
-        <enum>Qt::Horizontal</enum>
-       </property>
-       <property name="sizeHint" stdset="0">
-        <size>
-         <width>0</width>
-         <height>0</height>
-        </size>
-       </property>
-      </spacer>
-     </item>
-     <item>
-      <widget class="QPushButton" name="OK">
-       <property name="text">
-        <string>OK !</string>
-       </property>
-      </widget>
-     </item>
-     <item>
-      <spacer name="rightSpacer">
-       <property name="orientation">
-        <enum>Qt::Horizontal</enum>
-       </property>
-       <property name="sizeHint" stdset="0">
-        <size>
-         <width>0</width>
-         <height>0</height>
-        </size>
-       </property>
-      </spacer>
-     </item>
-    </layout>
-   </item>
-  </layout>
- </widget>
- <resources/>
- <connections/>
-</ui>
